Feature: Connect ToDo Page to DB
ToDo 페이지 DB연결,
ToDo 데이터베이스 나누어 작업,
ToDo 데이터 모델 나누어 작업,
API 서비스 나누어 작업,
군번 값을 static 값으로 전역으로 처리
