+++ lib/services/web_api/api_path.dart
@@ -1,0 +1,8 @@
+class APIPath {
+  static const String  origin = 'http://soldierank.shop/';
+  static const String  login = '${origin}get.php';
+  static const String  register = '${origin}register.php';
+  static const String  main = '${origin}getInfo2.php';
+  static const String  toDo = '${origin}todo.php';
+  static const String  group = '${origin}group.php';
+}
